+++ NEWS.md
@@ -8,7 +8,20 @@ Cite all releases with doi [10.5281/zenodo.5086371](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
-## IEATools 0.1.76 (2024-12-09)
+## IEATools 0.1.77 (2025-05-14)
+
+* New function `reallocate_statistical_differences()` reallocates
+  statistical differences in proportion to the non-zero consumption of each 
+  energy carrier in other industries.
+* `extend_to_useful()` now produces U_eiou_fu_details matrices with the
+  correct row and column name structure when there is no 
+  energy industry own use.
+* New tests for new features.
+    * Now at 1385 tests, all passing.
+    * Test coverage remains at 100%.
+
+
+## IEATools 0.1.76 (2024-12-09) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14335708.svg)](https://doi.org/10.5281/zenodo.14335708)
 
 * New function `load_electricity_heat_output()`
   creates a data frame of IEA electricity and heat output